Подключена библиотека FFTW. Производительность составила 1/5 от матлабовской. Дальнейшие оптимизации: использовать вещественное преобразование, убрать утечки памяти.

+++ NumericalMethod/minitypes.h
@@ -1,4 +1,4 @@
 #pragma once
 
-typedef float num;
+typedef double num;
 typedef unsigned int uint;
